Ready for ~pre2
git-svn-id: https://cppannotations.svn.sourceforge.net/svnroot/cppannotations/trunk@341 f6dd340e-d3f9-0310-b409-bdd246841980

+++ yo/pointermembers/examples/initializing.cc
@@ -1,19 +1,19 @@
     class PointerDemo
     {
         public:
-            unsigned d_value;
-            unsigned get() const;
+            size_t d_value;
+            size_t get() const;
     };
 
-    inline unsigned PointerDemo::get() const
+    inline size_t PointerDemo::get() const
     {
         return d_value;
     }
 
     int main()
     {                                           // initialization
-        unsigned (PointerDemo::*getPtr)() const = &PointerDemo::get;
-        unsigned PointerDemo::*valuePtr         = &PointerDemo::d_value;
+        size_t (PointerDemo::*getPtr)() const = &PointerDemo::get;
+        size_t PointerDemo::*valuePtr         = &PointerDemo::d_value;
 
         getPtr   = &PointerDemo::get;           // assignment
         valuePtr = &PointerDemo::d_value;
